LayoutTests/platform/blackberry/fast/dom/beforeload/pi-before-load-expected.txt

CONSOLE MESSAGE: line 35: NOT_SUPPORTED_ERR: DOM Exception 9: The implementation did not support the requested type of object or operation.
PASS
